refactor(storage): inline namespace and normalize into storage.ts
Consolidates the storage module by moving `namespace()` and
`normalizeKey()`/`normalizePrefix()` directly into `storage.ts`,
eliminating two single-purpose files. Also flows `evictAfterCycles`
to legacy offloader storage when the user hasn't explicitly set
`evictAfterTurns`.
