+++ strands-ts/src/storage/storage.ts
@@ -1,3 +1,40 @@
+import { StorageError } from '../errors.js'
+
+/**
+ * Validates and normalizes a storage key: collapses runs of `/`, strips leading
+ * and trailing `/`, and rejects empty keys and any `..` segment.
+ *
+ * @param key - The raw key to normalize
+ * @returns The normalized key
+ * @throws {@link StorageError} if the key is empty or contains a `..` segment
+ */
+export function normalizeKey(key: string): string {
+  const normalized = key.replace(/\/+/g, '/').replace(/^\/+|\/+$/g, '')
+  if (normalized.length === 0) {
+    throw new StorageError('Storage key must not be empty')
+  }
+  if (normalized.split('/').includes('..')) {
+    throw new StorageError(`Invalid storage key '${key}': '..' path segments are not allowed`)
+  }
+  return normalized
+}
+
+/**
+ * Normalizes a list prefix: collapses slash runs, strips leading slashes.
+ * Unlike a key, an empty prefix is valid and matches everything.
+ *
+ * @param prefix - The raw prefix to normalize
+ * @returns The normalized prefix
+ * @throws {@link StorageError} if the prefix contains a `..` segment
+ */
+export function normalizePrefix(prefix: string): string {
+  const normalized = prefix.replace(/\/+/g, '/').replace(/^\/+/, '')
+  if (normalized.split('/').includes('..')) {
+    throw new StorageError(`Invalid storage prefix '${prefix}': '..' path segments are not allowed`)
+  }
+  return normalized
+}
+
 /**
  * A backend for storing and retrieving raw bytes under string keys.
  *
@@ -55,4 +92,26 @@ export interface Storage<ListQuery = string> {
    * @throws {@link StorageError} if the listing fails
    */
   list(query: ListQuery): Promise<string[]>
+}
+
+/**
+ * Returns a {@link Storage} view with all keys prefixed by `prefix`.
+ *
+ * Composable — calling `namespace()` on the result nests prefixes.
+ *
+ * @internal
+ * @param storage - The underlying storage to delegate to
+ * @param prefix - Prefix to prepend to all keys
+ * @returns A namespaced Storage view
+ */
+export function namespace(storage: Storage, prefix: string): Storage & { namespace(prefix: string): Storage } {
+  const normalized = normalizePrefix(prefix)
+  const p = normalized ? `${normalized}/` : ''
+  return {
+    write: (key, data) => storage.write(`${p}${key}`, data),
+    read: (key) => storage.read(`${p}${key}`),
+    delete: (key) => storage.delete(`${p}${key}`),
+    list: (query) => storage.list(`${p}${query}`).then((keys) => keys.map((key) => key.slice(p.length))),
+    namespace: (sub) => namespace(storage, `${p}${sub}`),
+  }
 }
